Рефакторинг, убрал не используемый код, поправил предупреждения от intelji

+++ src/dimant/simulation/service/CreatureService.java
@@ -12,9 +12,9 @@ import dimant.simulation.utils.RandomIntValue;
 import java.util.List;
 
 public class CreatureService {
-    private MapBoard mapBoard;
-    private Navigator navigator;
-    private EntityService entityService;
+    private final MapBoard mapBoard;
+    private final Navigator navigator;
+    private final EntityService entityService;
 
     public CreatureService(MapBoard mapBoard, Navigator navigator, EntityService entityService) {
         this.mapBoard = mapBoard;
@@ -30,10 +30,8 @@ public class CreatureService {
             if (mapBoard.hasEntityOnMapBoard(creature)) {
                 String typeNextMove = navigator.getTypeNextStep(creature);
                 triggerActiveCreature(creature, typeNextMove);
-                creatures.remove(index);
-            } else {
-                creatures.remove(index);
             }
+            creatures.remove(index);
         }
     }
 
@@ -42,26 +40,13 @@ public class CreatureService {
         System.out.println("сходил " + creature.toString());
         switch (reaction) {
             case GO -> moveCreature(creature);
-            case ATTACK -> attackHerbivore(creature); // для wolf когда makeMove вернул Hare
-            case EAT -> eatGrass(); // для hare когда makeMove вернул Grass
+            case ATTACK -> attackHerbivore(creature);
+            case EAT -> eatGrass();
             case GO_GRASS -> goGrass(creature);
             default -> System.out.println("Действие не определено" + creature.toString()); // xз чё тут сделать, может остаться на месте
         }
-        ;
     }
 
-//    public void moveCreature(Creature creature) {
-//        Coordinate currentCoordinate = mapBoard.getCreatureCoordinate(creature);
-//        Coordinate nextCoordinate = navigator.getNextCoordinateCreature();
-//        if (entityService.hasGrassByCoordinateInMapGrass(currentCoordinate)) {
-//            Grass grass = entityService.getGrassMapGrass(currentCoordinate);
-//            mapBoard.addEntityMap(currentCoordinate, grass);
-//            mapBoard.addEntityMap(nextCoordinate, creature);
-//            return;
-//        }
-//        mapBoard.addEntityMap(nextCoordinate, creature);
-//        mapBoard.addEntityMap(currentCoordinate, new Place());
-//    }
 
     public void moveCreature(Creature creature) {
         Coordinate currentCoordinate = mapBoard.getCreatureCoordinate(creature);
@@ -74,14 +59,13 @@ public class CreatureService {
     public void attackHerbivore(Creature creature) {
         Coordinate nextCoordinate = navigator.getNextCoordinateCreature();
         Hare hare = mapBoard.getHare(nextCoordinate);
-        creature.attack(hare);
-        if (isHPHareLow(hare)) {
-            entityService.deleteEntityMap(hare);
+        if (hare != null) {
+            creature.attack(hare);
+            if (isHPHareLow(hare)) {
+                entityService.deleteEntityMap(hare);
+            }
         }
     }
-
-    // TODO: 06.11.2024 Не совсем корректно сюда добавлять проверку на уровень hp и удаления hare с карты. Скорее всего после каждого хода entityService должен производить свою работу по удалению и добавлению сущностей
-
 
     public void eatGrass() {
         Coordinate nextCoordinate = navigator.getNextCoordinateCreature();
@@ -103,6 +87,4 @@ public class CreatureService {
     private boolean isHPHareLow(Hare hare) {
         return hare.getHP() <= 0;
     }
-
-
 }
